+++ src/navigation/AppNavigator.tsx
@@ -25,7 +25,6 @@ const TabNavigator = () => {
       <Tab.Screen name="Invest" component={InvestScreen} />
       <Tab.Screen name="Analysis" component={AnalysisScreen} />
       <Tab.Screen name="Bag" component={BagScreen} />
-      <Tab.Screen name="Transaction" component={TransactionScreen} />
     </Tab.Navigator>
   );
 };
@@ -34,13 +33,8 @@ const AppNavigator = () => {
   return (
     <Stack.Navigator screenOptions={{headerShown: false}}>
       <Stack.Screen name="MainTabs" component={TabNavigator} />
-      <Stack.Screen
-        name="Settings"
-        component={SettingsScreen}
-        options={{
-          presentation: 'modal',
-        }}
-      />
+      <Stack.Screen name="Settings" component={SettingsScreen} options={{ presentation: 'modal',}}/>
+      <Stack.Screen name="Transaction" component={TransactionScreen} options={{ presentation: 'modal' }} />
     </Stack.Navigator>
   );
 };
